Add back derive_has_field_impls_from_struct
