+++ crates/macros/cgp-macro-core/src/types/cgp_data/mod.rs
@@ -1,9 +1,11 @@
 mod derive_builder;
 mod derive_extractor;
+mod derive_has_field;
 mod derive_has_fields;
 mod record;
 
 pub use derive_builder::*;
 pub use derive_extractor::*;
+pub use derive_has_field::*;
 pub use derive_has_fields::*;
 pub use record::*;
